requirements_modular.txt
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
PyPDF2>=3.0.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain>=0.0.300
transformers>=4.30.0
openai>=1.0.0
cerebras-cloud-sdk>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=2.0.0 